Reorganization
Arnoldi Module --> dynamic library

+++ scilab/modules/arnoldi/sci_gateway/gw_arnoldi.c
@@ -6,6 +6,7 @@
 /*-----------------------------------------------------------------------------------*/
 #include "MALLOC.h"
 #include "stack-c.h"
+#include "sciprint.h"
 /*-----------------------------------------------------------------------------------*/
 #if _MSC_VER
 extern char *GetExceptionString(DWORD ExceptionCode);
@@ -43,7 +44,7 @@ int C2F(gw_arnoldi)()
 		#ifndef _DEBUG
 		_try
 		{
-			(*(Tab[Fin-1].f)) (Tab[Fin-1].name,strlen(Tab[Fin-1].name));
+			(*(Tab[Fin-1].f)) (Tab[Fin-1].name,(unsigned long)strlen(Tab[Fin-1].name));
 		}
 		_except (EXCEPTION_EXECUTE_HANDLER)
 		{
@@ -52,10 +53,10 @@ int C2F(gw_arnoldi)()
 			if (ExceptionString) {FREE(ExceptionString);ExceptionString=NULL;}
 		}
 		#else
-			(*(Tab[Fin-1].f)) (Tab[Fin-1].name,strlen(Tab[Fin-1].name));
+			(*(Tab[Fin-1].f)) (Tab[Fin-1].name,(unsigned long)strlen(Tab[Fin-1].name));
 		#endif
 	#else
-		(*(Tab[Fin-1].f)) (Tab[Fin-1].name,strlen(Tab[Fin-1].name));
+		(*(Tab[Fin-1].f)) (Tab[Fin-1].name,(unsigned long)strlen(Tab[Fin-1].name));
 	#endif
 
 	return 0;
